requirements.txt
pytest>=8.2
pytest-xdist>=3.6
pytest-rerunfailures>=14.0

playwright>=1.45

requests>=2.32
pydantic>=2.8

python-dotenv>=1.0

black>=24.8
ruff>=0.6
isort>=6.14.5
mypy>=1.4.1